refactor: move EnsureActiveLocationInitializedUseCase to subpackage ensure_active_location
